refactor: product detail 관련 components는 [id] 안에
