+++ app/product/[id]/components/AddToCartForm.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import ProductOptions from "@/app/product/components/ProductOptions";
-import ProductQuantity from "@/app/product/components/ProductQuantity";
-import useAddToCartForm from "@/app/product/hooks/forms/useAddToCartForm";
+import ProductOptions from "@/app/product/[id]/components/ProductOptions";
+import ProductQuantity from "@/app/product/[id]/components/ProductQuantity";
+import useAddToCartForm from "@/app/product/[id]/hooks/forms/useAddToCartForm";
 import { Button } from "@/components/ui/button";
 import { Form } from "@/components/ui/form";
 import { ProductDetailInfo } from "@/lib/types/productType";
