+++ app/product/[id]/components/ProductDetail.tsx
