+++ app/product/[id]/components/ProductNavbar.tsx
