+++ app/product/[id]/hooks/useProductNavigation.tsx
